Добавляет задачу 5 Модуляция
