+++ index.html
@@ -231,6 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-<script src="./js/main.js"></script>
+<script type="module" src="./js/main.js"></script>
 </body>
 </html>
