+++ js/create_id_generator.js
@@ -1,0 +1,6 @@
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+  return () => (lastGeneratedId += 1);
+};
+
+export { createIdGenerator };
